==> 1. get rdkit alogp.ipynb <==
import pandas as pd
import numpy as np
import rdkit as rk
data1 = pd.read_csv('0. input data.csv')
data2 = data1.loc[:, ["SMILES", "Kow", "Kowwin"]]

from rdkit.Chem import QED
data2['Wildman-Crippen LogP'] = data2['SMILES'].apply(lambda smi: rk.Chem.MolFromSmiles(smi)).apply(lambda mol: rk.Chem.Crippen.MolLogP(mol) if mol else np.nan)

data2.to_csv('1.1. data with Wildman-Crippen.csv', index=False)
