==> 4. unify smi strings for vega output.ipynb <==
import pandas as pd
import numpy as np
import rdkit as rk
data = pd.read_csv("2.2. vega output.csv", header=6).drop(['No.','tId','LogP model (Meylan-Kowwin)-assessment','LogP model (MLogP)-assessment','LogP model (ALogP)-assessment'], axis=1)

from rdkit import Chem
data['SMILES'] = data['SMILES'].apply(lambda smi: Chem.MolFromSmiles(smi)).apply(lambda mol: Chem.MolToSmiles(mol, isomericSmiles=False, canonical=True) if mol else None)

data.rename(columns={'SMILES': 'SMILES2'}, inplace=True)

data.to_csv("4.1. vega output cleaned.csv", index=False)
