==> 5. merge data and remove undefined smi.ipynb <==
import pandas as pd
import numpy as np
import rdkit as rk
data1 = pd.read_csv("3.1. data with alogp cleaned.csv")
data2 = pd.read_csv("4.1. vega output cleaned.csv")

data = data1.join(data2).drop(columns=['SMILES2'])

data['LogP model (ALogP)-prediction'] = data['LogP model (ALogP)-prediction'].replace('-', np.nan)
data['LogP model (MLogP)-prediction'] = data['LogP model (MLogP)-prediction'].replace('-', np.nan)
data['LogP model (Meylan-Kowwin)-prediction'] = data['LogP model (Meylan-Kowwin)-prediction'].replace('-', np.nan)

data['defined smi'] = data['SMILES'].apply(lambda smi: smi is not np.nan)
data = data[data['defined smi'] == True].drop(columns=['defined smi'])

data.to_csv("5.1. cleaned and merged data.csv", index=False)
