==> 6. append molecular features.ipynb <==
import rdkit as rk
import pandas as pd
import numpy as np
data = pd.read_csv('5.1. cleaned and merged data.csv')

from rdkit import Chem
from rdkit.Chem import Fragments
from rdkit.Chem import Lipinski
data['mols'] = data['SMILES'].apply(lambda smi: Chem.MolFromSmiles(smi))
data['carboxylic acid'] = data['mols'].apply(lambda mol: Fragments.fr_COO2(mol))
data['amide'] = data['mols'].apply(lambda mol: Fragments.fr_amide(mol))
data['ester'] = data['mols'].apply(lambda mol: Fragments.fr_ester(mol))
data['benzene'] = data['mols'].apply(lambda mol: Fragments.fr_benzene(mol))
data['ketone'] = data['mols'].apply(lambda mol: Fragments.fr_ketone(mol))
data['halogen'] = data['mols'].apply(lambda mol: Fragments.fr_halogen(mol))
data['primary amine'] = data['mols'].apply(lambda mol: Fragments.fr_NH2(mol))
data['secondary amine'] = data['mols'].apply(lambda mol: Fragments.fr_NH1(mol))
data['tertiary amine'] = data['mols'].apply(lambda mol: Fragments.fr_NH0(mol))
data['aliphatic hydroxyl'] = data['mols'].apply(lambda mol: Fragments.fr_Al_OH(mol))
data['unbranched alkane >4'] = data['mols'].apply(lambda mol: Fragments.fr_unbrch_alkane(mol))
data['alkyl halide'] = data['mols'].apply(lambda mol: Fragments.fr_alkyl_halide(mol))
data['nitrile'] = data['mols'].apply(lambda mol: Fragments.fr_nitrile(mol))
data['nitro'] = data['mols'].apply(lambda mol: Fragments.fr_nitro(mol))
data['sulfide'] = data['mols'].apply(lambda mol: Fragments.fr_sulfide(mol))
data['sulfonamide'] = data['mols'].apply(lambda mol: Fragments.fr_sulfonamd(mol))
data['ether'] = data['mols'].apply(lambda mol: Fragments.fr_ether(mol))
data['aromatic hydroxyl'] = data['mols'].apply(lambda mol: Fragments.fr_Ar_OH(mol))
data['aromatic nitrogen'] = data['mols'].apply(lambda mol: Fragments.fr_Ar_N(mol))
data['aromatic amine'] = data['mols'].apply(lambda mol: Fragments.fr_Ar_NH(mol))
data['aniline'] = data['mols'].apply(lambda mol: Fragments.fr_aniline(mol))
data['phenol'] = data['mols'].apply(lambda mol: Fragments.fr_phenol(mol))
data['bicyclic'] = data['mols'].apply(lambda mol: Fragments.fr_bicyclic(mol))
data['rings'] = data['mols'].apply(lambda mol: Lipinski.RingCount(mol))
data['saturated rings'] = data['mols'].apply(lambda mol: Lipinski.NumSaturatedRings(mol))
data['aromatic rings'] = data['mols'].apply(lambda mol: Lipinski.NumAromaticRings(mol))

data.to_csv('6.1. data with molecular features.csv', index=False)
