==> 6.2. get molecular property counts.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
data = pd.read_csv('7.1. data with error values.csv')
data1 = data.drop(columns=data.columns[0:8]).drop(columns=data.columns[34])

propertylist = data1.columns[0:26]
n = 0
while n < len(propertylist):
    data1[propertylist[n]] = data1[propertylist[n]].apply(lambda prop: prop > 0)
    n += 1

data2 = pd.DataFrame(columns=propertylist, index=['True', 'False'])
for prop in propertylist:
    data2.loc['True', prop] = data1[prop].sum()
    data2.loc['False', prop] = len(data1) - data1[prop].sum()
data2 = data2.transpose(copy=True)
data2.sort_values(by='True', inplace=True)

data2.to_csv('6.2.1. molecular property counts.csv')
