==> 7. append error values.ipynb <==
import numpy as np
import pandas as pd
data = pd.read_csv('6.1. data with molecular features.csv')

data['consensus logp'] = data[['Kowwin', 'Wildman-Crippen LogP', 'LogP model (Meylan-Kowwin)-prediction', 'LogP model (MLogP)-prediction', 'LogP model (ALogP)-prediction']].mean(axis=1)
data['Kowwin error'] = (data['Kowwin'] - data['Kow'])
data['Wildman-Crippen error'] = (data['Wildman-Crippen LogP'] - data['Kow'])
data['Meylan-Kowwin error'] = (data['LogP model (Meylan-Kowwin)-prediction'] - data['Kow'])
data['MLogP error'] = (data['LogP model (MLogP)-prediction'] - data['Kow'])
data['ALogP error'] = (data['LogP model (ALogP)-prediction'] - data['Kow'])
data['consensus logp error'] = (data['consensus logp'] - data['Kow'])

data['Abs Kowwin error'] = np.abs(data['Kowwin'] - data['Kow'])
data['Abs Wildman-Crippen error'] = np.abs(data['Wildman-Crippen LogP'] - data['Kow'])
data['Abs Meylan-Kowwin error'] = np.abs(data['LogP model (Meylan-Kowwin)-prediction'] - data['Kow'])
data['Abs MLogP error'] = np.abs(data['LogP model (MLogP)-prediction'] - data['Kow'])
data['Abs ALogP error'] = np.abs(data['LogP model (ALogP)-prediction'] - data['Kow'])
data['Abs consensus logp error'] = np.abs(data['consensus logp'] - data['Kow'])

data.to_csv('7.1. data with error values.csv', index=False)
