==> 8. statistical testing.ipynb <==
import numpy as np
import scipy as sp
import pandas as pd
data = pd.read_csv('7.1. data with error values.csv')

def compounds_with(column, feature):
    return data[column].loc[data[feature] > 0]
def compounds_without(column, feature):
    return data[column].loc[data[feature] == 0]

modellist = data.loc[:, 'Abs Kowwin error':'Abs consensus logp error'].columns.values
propertylist = data.loc[:, 'carboxylic acid':'aromatic rings'].columns.values
header = ['Statistic', 'P-Value', 'Degrees of Freedom']
ind = []
propertywrite = 0
while propertywrite < len(propertylist):
    propertyx = propertylist[propertywrite]
    modelwrite = 0
    while modelwrite < len(modellist):
        modelx = modellist[modelwrite]
        ind.append(propertyx + ', ' + modelx)
        modelwrite += 1
    propertywrite += 1
data1 = pd.DataFrame(columns=header, index=pd.Index(ind, name='Model and Property'))

from scipy import stats
modelrun = 0
while modelrun < len(modellist):
    propertyrun = 0
    while propertyrun < len(propertylist):
        model = modellist[modelrun]
        property = propertylist[propertyrun]
        x = compounds_with(model, property)
        y = compounds_without(model, property)
        a = [(stats.ttest_ind(x, y, equal_var=False, nan_policy='omit').statistic),
            (stats.ttest_ind(x, y, equal_var=False, nan_policy='omit').pvalue),
            (stats.ttest_ind(x, y, equal_var=False, nan_policy='omit').df),
        ]
        n = 0
        while n < len(a):
            data1.loc[(property + ', ' + model), (header[n])] = a[n]
            n += 1
        propertyrun += 1
    modelrun += 1

data1.to_csv('8.1. statistical testing results.csv')
