==> 9. analysis of statistical results.ipynb <==
import pandas as pd
import numpy as np
mols = pd.read_csv('7.1. data with error values.csv')
statstest = pd.read_csv('8.1. statistical testing results.csv')

errorstats = pd.DataFrame(index = mols.loc[:, 'Abs Kowwin error':'Abs consensus logp error'].columns.values, columns=['Mean', 'Standard Deviation', 'Minimum', 'Maximum', 'Median', 'IQR'])
errorstats['Mean'] = mols.loc[:, 'Abs Kowwin error':'Abs consensus logp error'].mean(axis=0).round(5)
errorstats['Standard Deviation'] = mols.loc[:, 'Abs Kowwin error':'Abs consensus logp error'].std(axis=0).round(5)
errorstats['Minimum'] = mols.loc[:, 'Abs Kowwin error':'Abs consensus logp error'].min(axis=0).round(4)
errorstats['Maximum'] = mols.loc[:, 'Abs Kowwin error':'Abs consensus logp error'].max(axis=0).round(4)
errorstats['Median'] = mols.loc[:, 'Abs Kowwin error':'Abs consensus logp error'].median(axis=0).round(5)
errorstats['IQR'] = (mols.loc[:, 'Abs Kowwin error':'Abs consensus logp error'].quantile(0.75) - mols.loc[:, 'Abs Kowwin error':'Abs consensus logp error'].quantile(0.25)).round(5)
errorstats.to_csv('9.1. error statistics.csv')

def compounds_with(column, feature):
    return mols[column].loc[mols[feature] > 0]
def compounds_without(column, feature):
    return mols[column].loc[mols[feature] == 0]

import re
features = mols.loc[:, 'carboxylic acid':'aromatic rings'].columns.values
methods = mols.loc[:, 'Abs Kowwin error':'Abs consensus logp error'].columns.values.tolist()
methods = [re.sub('Abs ', '', method) for method in methods]
methods = [re.sub(' error', '', method) for method in methods]
index = []
cols = ['Count']
for feature in features:
    index.append('With ' + feature)
    index.append('Without ' + feature)
for method in methods:
    cols.append(method + ' Mean')
    cols.append(method + ' Standard Deviation')
    cols.append(method + ' Minimum')
    cols.append(method + ' Maximum')
    cols.append(method + ' Median')
    cols.append(method + ' Q1')
    cols.append(method + ' Q3')
data = pd.DataFrame(index = index, columns = cols)
for feature in features:
    data.loc[('With ' + feature), 'Count'] = len(compounds_with('Abs Kowwin error', feature))
    data.loc[('Without ' + feature), 'Count'] = len(compounds_without('Abs Kowwin error', feature))
    for method in methods:
        data.loc[('With ' + feature), (method + ' Mean')] = compounds_with('Abs ' + method + ' error', feature).mean()
        data.loc[('Without ' + feature), (method + ' Mean')] = compounds_without('Abs ' + method + ' error', feature).mean()
        data.loc[('With ' + feature), (method + ' Standard Deviation')] = compounds_with('Abs ' + method + ' error', feature).std()
        data.loc[('Without ' + feature), (method + ' Standard Deviation')] = compounds_without('Abs ' + method + ' error', feature).std()
        data.loc[('With ' + feature), (method + ' Minimum')] = compounds_with('Abs ' + method + ' error', feature).min()
        data.loc[('Without ' + feature), (method + ' Minimum')] = compounds_without('Abs ' + method + ' error', feature).min()
        data.loc[('With ' + feature), (method + ' Maximum')] = compounds_with('Abs ' + method + ' error', feature).max()
        data.loc[('Without ' + feature), (method + ' Maximum')] = compounds_without('Abs ' + method + ' error', feature).max()
        data.loc[('With ' + feature), (method + ' Median')] = compounds_with('Abs ' + method + ' error', feature).median()
        data.loc[('Without ' + feature), (method + ' Median')] = compounds_without('Abs ' + method + ' error', feature).median()
        data.loc[('With ' + feature), (method + ' Q1')] = (compounds_with('Abs ' + method + ' error', feature).quantile(0.25))
        data.loc[('Without ' + feature), (method + ' Q1')] = (compounds_without('Abs ' + method + ' error', feature).quantile(0.25))
        data.loc[('With ' + feature), (method + ' Q3')] = (compounds_with('Abs ' + method + ' error', feature).quantile(0.75))
        data.loc[('Without ' + feature), (method + ' Q3')] = (compounds_without('Abs ' + method + ' error', feature).quantile(0.75))
data.to_csv('9.2. error statistics by feature.csv')
