==> 10. making the stats into nice tables.ipynb <==
import pandas as pd
import numpy as np
import re
stats = pd.read_csv('9.3. error statistics by feature with excel stuff.csv', index_col=0)
stats2 = pd.read_csv('8.1. statistical testing results.csv', index_col=0).transpose()

props = stats[::2].index.values
props = [re.sub('With ', '', prop) for prop in props]
vals = stats.loc[:, 'Kowwin Mean':'consensus logp IQR'].columns.values
methods = vals[::6]
methods = [re.sub(' Mean', '', method) for method in methods]
valsnew = []
loop = 0
while loop < (len(vals) / 6):
    valsnew.append((methods[loop] + ' T-statistic'))
    valsnew.append((methods[loop] + ' p-value'))
    valsnew.append((methods[loop] + ' Degrees of Freedom'))
    valsnew.append(vals[loop * 6])
    valsnew.append(vals[loop * 6 + 1])
    valsnew.append(vals[loop * 6 + 3])
    valsnew.append(vals[loop * 6 + 4])
    valsnew.append(vals[loop * 6 + 5])
    loop += 1
propsnew = stats.index.values
propsnew = [re.sub('With ', '', prop) for prop in propsnew]
propsnew = [re.sub('Without', 'No', prop) for prop in propsnew]
full_stats = pd.DataFrame(index=propsnew, columns=valsnew)

for method in methods:
    for prop in props:
        full_stats.loc[prop, method + ' T-statistic'] = stats2.loc['Statistic', prop + ', Abs ' + method + ' error']
        full_stats.loc[prop, method + ' p-value'] = stats2.loc['P-Value', prop + ', Abs ' + method + ' error']
        full_stats.loc[prop, method + ' Degrees of Freedom'] = stats2.loc['Degrees of Freedom', prop + ', Abs ' + method + ' error']

for val in vals:
    for prop in props:
        full_stats.loc[prop, val] = stats.loc['With ' + prop, val]
        full_stats.loc['No ' + prop, val] = stats.loc['Without ' + prop, val]
full_stats.to_csv('10.1. full error statistics and statistical testing.csv')
